requirements.txt
streamlit==1.37.1
pybaseball==2.2.7
pandas==2.2.2
plotly==5.24.1
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pyarrow==16.1.0
scipy==1.13.1
matplotlib==3.9.2
tqdm==4.66.5
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
scikit-learn==1.3.2
seaborn==0.13.0
MLB-StatsAPI==1.6.2